Add a function for simple XPath evaluation.
The existing `evaluate` function requires lots of research,
configuration and additional imports to use, while the poor souls that
have to process XML usually just need to be able to execute simple
XPath expressions. The `evaluate_xpath` function only needs a
`Document` and an XPath `&str` to work and it returns a `Value` in
wrappers ready to accomodate various possible errors.