Add a function for simple, absolute XPath evaluation. The existing `evaluate` function requires lots of research, configuration and additional imports to use, while the poor souls that have to process XMLs usually just need to be able to execute simple XPath expressions. The `evaluate_absolute_xpath` function only needs a `Document` and an XPath `&str` to work and it returns a `Value` in wrappers ready to accomodate various possible errors.
In addition, this extends the `expression::Error`struct with an `InvalidXPath` variant so that potential XPath-related `parser::ParseErr` errors can be handled. It is not perfect; instead of a `String` I would prefer to relay `ParseErr`, but it would break the `ExpressionError`'s `Hash` property - deriving `Hash` would fail due to `Token` having an `f64` variant.

The last modification in the `integration.rs` file is just a deletion of an unused trait import (`Expression`).